move getAuthorizedUserPublicKey into auth.go
