+++ server/secretshandler.go
@@ -110,44 +110,6 @@ func encryptSecretMetadata(metadata v1structs.SecretMetadata, key *pgpkey.PgpKey
 	return encrypted, nil
 }
 
-func getAuthorizedUserPublicKey(r *http.Request) (*pgpkey.PgpKey, error) {
-	// TODO: actually authenticate a public key!
-	//
-	// For now anyone can "authenticate" as any public key which is
-	// obviously stupid, but the impact is limited by the fact that all
-	// content is encrypted to the public key.
-	//
-	// Look for a header like:
-	// Authorization: tmpfingerprint: OPENPGP4FPR:AAAABBBBAAAABBBBAAAABBBBAAAABBBBAAAABBBB
-
-	const prefix string = "tmpfingerprint: OPENPGP4FPR:"
-
-	authHeader := r.Header.Get("Authorization")
-	if !strings.HasPrefix(authHeader, prefix) {
-		return nil, fmt.Errorf("missing Authorization header starting `tmpfingerprint: OPENPGP4FPR:`")
-	}
-
-	fpr, err := fingerprint.Parse(authHeader[len(prefix):])
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse fingerprint: %v", err)
-	}
-
-	armoredPublicKey, found, err := datastore.GetArmoredPublicKeyForFingerprint(fpr)
-	if err != nil {
-		return nil, err
-	} else if !found {
-		return nil, fmt.Errorf("invalid authorization")
-	}
-
-	key, err := pgpkey.LoadFromArmoredPublicKey(armoredPublicKey)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to load key: %v", err)
-	}
-
-	return key, nil
-}
-
 func parseFingerprint(fp string) (*fingerprint.Fingerprint, error) {
 	if !strings.HasPrefix(fp, "OPENPGP4FPR:") {
 		return nil, fmt.Errorf("missing prefix `OPENPGP4FPR:`")
